+++ src/main/java/de/samply/frontend/FrontendService.java
@@ -9,6 +9,7 @@ import de.samply.app.ProjectManagerController;
 import de.samply.user.roles.RolesExtractor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.web.util.UriComponentsBuilder;
 
 import java.util.*;
 
@@ -16,9 +17,13 @@ import java.util.*;
 public class FrontendService {
 
     private final ConstraintsService constraintsService;
+    private final FrontendConfiguration frontendConfiguration;
 
-    public FrontendService(ConstraintsService constraintsService) {
+    public FrontendService(
+            ConstraintsService constraintsService,
+            FrontendConfiguration frontendConfiguration) {
         this.constraintsService = constraintsService;
+        this.frontendConfiguration = frontendConfiguration;
     }
 
     public Collection<ModuleActionsPackage> fetchModuleActionPackage(String site, Optional<String> projectCode, Optional<String> bridgehead) {
@@ -48,6 +53,21 @@ public class FrontendService {
             }
         });
         return moduleModuleActionsPackageMap.values();
+    }
+
+    public String fetchUrl(String site, Map<String, String> parameters) {
+        UriComponentsBuilder result = UriComponentsBuilder.fromHttpUrl(frontendConfiguration.getBaseUrl());
+        if (site != null) {
+            Optional<String> sitePath = frontendConfiguration.getSitePath(site);
+            if (sitePath.isPresent()) {
+                result.path(sitePath.get());
+            }
+        }
+        if (parameters != null && !parameters.isEmpty()) {
+            parameters.keySet().forEach(parameter ->
+                    result.queryParamIfPresent(parameter, Optional.ofNullable(parameters.get(parameter))));
+        }
+        return result.toUriString();
     }
 
 }